diffusion: SequenceParallelPlan — per-family SP declaration behind ModelBackend
A model family opts into SP by declaring a SequenceParallelPlan (boundaries,
attention installer, head count) instead of the engine hard-coding Wan:

- ModelBackend.sequence_parallel_plan assembles the plan; the diffusers
  default reads the model's _cp_plan and the family config's
  apply_sp_attention hook. Native backends override the method wholesale.
- WanUSPAttnProcessor moves to configs/wan2_2.py; the engine layer
  (sp_attention/sp_ops/sp_mesh/parallel) no longer references any model class.
- The isinstance whitelist becomes capability checks: no _cp_plan or no
  apply_sp_attention raises with the missing declaration named.

Validated: CPU suites (88 passed) + full SP GPU suite (21 bands, 66 checks).
