+++ miles/backends/fsdp_utils/configs/train_pipeline_config.py
@@ -184,6 +184,14 @@ class TrainPipelineConfig(abc.ABC):
     ) -> torch.Tensor:
         """Apply classifier-free guidance. Model-specific (e.g. rescale or not)."""
 
+    @abc.abstractmethod
+    def preprocess_model_before_fsdp(self, model: torch.nn.Module) -> None:
+        """Preprocess the model before FSDP."""
+
+    def apply_sp_attention(self, model: torch.nn.Module, parallel_state) -> None:
+        """Route this family's self-attention through ``sp_ops.usp_attention``."""
+        raise NotImplementedError(f"{type(self).__name__} does not implement sequence-parallel attention")
+
     def postprocess_model_after_materialize(self, model: torch.nn.Module) -> None:
         """Postprocess the model after FSDP wrap + weight materialization (default: no-op)."""
         return None
